Fixes #204

+++ src/Controllers/DashSettingsController.php
@@ -113,7 +113,7 @@ class DashSettingsController extends Controller
      */
     public function postSettings()
     {
-        if (Input::get('remove_banner') == "1") {
+        if (Input::has('remove_banner') && Input::get('remove_banner') == "1") {
             $setting = Setting::where('name', 'app_banner');
             $setting->delete();
         }
